Update html and MathJax for Lighthouse 100% Score
Unfortunately this not only defers the loading of MathJax,
but also makes the extension load the plugins individually.
This is to reduce the amount of Javascript that needs to
be loaded, but does add extra network requests. Lighthouse
doesn't like what the implementation was before and even
still it only reports a 98% score on those pages (though
when deployed it may be higher due to gzip compression).
It may be noted that MathJax is only loaded on blog posts
that explcitly require it, thus the non-100% score only
affects those blog posts.