Improve accessibility of <details> text
It appears that there are two issues with using
the `<details>` tag in HTML. (1) is that it can
be a problem for people with accessibility
technologies such as screenreaders. (2) The
latest version of IE (before Edge), does not
support the tag, and so it's best to fallback on
having a header tag.