Satisfy Lighthouse complaint about image res.
Lighthouse was complaining that I was serving a
320x320 px image, but it wanted a 640x640 px image.
I believe that this has to do with a 2x pixel density
on some devices. To satify it, I added the 2x, 640x640
version of my avatar.

Also I change the `width` and `height` attributes of the
avatar from 350px to 320px. I'm not quite sure why I
had it set to that. Perhaps it was a typo.