Convert from MathJAX to client-side KaTeX
Before I had been using a MathJAX script that was local to the hosted
HTML. I had done this out of a concern for privacy, but I'm not sure how
important that is in practice -- especially with the
`crossorigin="anonymous"` HTML attribute.

This also adds some extra Jekyll plugins for later.