Merge `background-color` into `background`
It appears that on some browsers, the `background-color: black` would be
overridden by `background: url(...)`. I observed this being a problem
when I left my website on during the light-dark nightly transition. It
seemed as though the dark background image did not load, and so the
website because a blinding white (since the `background-color` was not
set correctly).

This commit also changes paths to background images referenced in the
SASS from being relative (to what???) to absolute.