Add navigation bar to Blog and Bookshelf pages
The primary objective of this commit is to make navigation more
consistent and obvious to the user across the site. Previously it was
extremely unclear how to navigate outside the blog and bookshelf area,
as the user was expected to click on the title of the page.

The navigation bar from the Home page is now carried between pages,
however the anchor links from the Home page have been replaced with a
single "Home" link (as would be expected due to the locality of the
anchor links).

The heading tags (h1-h6) have been (more) standardized, which I believe
is important for people using assistive technologies like screen
readers. Before it was a bit haphazard, but now that the header of the
blog page/posts is more standard, the intent is clearer to the writer.