Replace manual permalinks in posts with global val
Before there was a manual permalink specified in each blog post that
placed it into the `/blog` path, as well under the generic
`/blog/posts/` sub-path (which ignores the category of the post).

This was simplified by using the `permalink` key in `_config.yml` that
can be set to a pattern to generate from.