Migrate a[href] from text-decoration to border
Safari has /interesting/ support in the realm of text-decoration. A
browser-specific prefix has to be used, and the line weight needs to be
in a separate tag.

Even when this patch is applied, there are problems with "descenders"
making the link look gross.

The solution I came up with is to just migrate to using border-bottom.
So far the fix looks like it has applied well.