Battle >> some more
The ">>" indicator for external links has always been troublesome. I
thought I had fixed this before, but when the browser's width is not
correct, ">>" could word-wrap away from the link itself (which is ugly).

Additionally there always seems to be problems with the ">>" being
properly aligned to the link itself. Sometimes the problem is more
obvious than other times.

Part of the problem stems from issues with `vertical-align`, but this is
compounded by the non-`1.0em` font-size I have been using. I believe
this was a stylistic choice, but it makes vertical alignment extremely
hard due to the different font-sizes.

The fix I've applied is to assume `vertical-align: baseline` as well as
`display: inline` and removing `font-size`. Hopefully this more basic
approach prevents further issues. Additionally it appears that the
`font-size` that was chosen is not really an issue stylistically.