Fix ambiguous resume anchor link
I myself thought that the "you can find my email in my resume" link at
the bottom of the home page was confusing, since it would take you to
the anchor like right above it (which often did not cause the page to
change). It is much simplier to just link directly to my resume. For
this I've added the resume link to `_config.yml` for DRY.