Fix Atom feed bug caused by _config.yml
I created a duplicate key in _config.yml -- the `author` key -- which
was used by jekyll-feed to grab the authors name. As a result, instead
of my name appear in the Atom feed, there was a serialized version of my
social media accounts, since I began to use that key for my social
feeds.