Update site news
I am searching for a job right now, and it might be a good idea to put
that in the "Recent News" section!

I also updated the contact page to put the "Platforms" section above the
"Coffee Chat" section. It is likely more important to people that they
are able to contact me rather than schedule a "coffee chat".